remove old code, move new code to its own file
